rip out old chunks parser
