+++ core/src/parsers/pyreport/mod.rs
@@ -1,15 +1,10 @@
 use std::fs::File;
 
 use memmap2::Mmap;
-use winnow::Parser;
 
-use crate::{
-    error::{CodecovError, Result},
-    report::{SqliteReport, SqliteReportBuilder, SqliteReportBuilderTx},
-};
+use crate::{error::Result, report::SqliteReportBuilder};
 
 pub mod chunks;
-pub mod chunks_serde;
 pub mod report_json;
 
 mod utils;
@@ -40,34 +35,16 @@ pub fn parse_pyreport(
     chunks_file: &File,
     report_builder: &mut SqliteReportBuilder,
 ) -> Result<()> {
-    // Encapsulate all of this in a block so that `report_builder_tx` gets torn down
-    // at the end. Otherwise, it'll hold onto a reference to `report_builder`
-    // and prevent us from consuming `report_builder` to actually build a
-    // `SqliteReport`.
-    {
-        let mut report_builder_tx = report_builder.transaction()?;
+    let mut report_builder_tx = report_builder.transaction()?;
 
-        // Memory-map the input file so we don't have to read the whole thing into RAM
-        let mmap_handle = unsafe { Mmap::map(report_json_file)? };
-        let report_json::ParsedReportJson { files, sessions } =
-            report_json::parse_report_json(&mmap_handle, &mut report_builder_tx)?;
+    // Memory-map the input file so we don't have to read the whole thing into RAM
+    let report_json_file = unsafe { Mmap::map(report_json_file)? };
+    let report_json = report_json::parse_report_json(&report_json_file, &mut report_builder_tx)?;
 
-        // Replace our mmap handle so the first one can be unmapped
-        let mmap_handle = unsafe { Mmap::map(chunks_file)? };
-        let buf = unsafe { std::str::from_utf8_unchecked(&mmap_handle[..]) };
+    // Replace our mmap handle so the first one can be unmapped
+    let chunks_file = unsafe { Mmap::map(chunks_file)? };
 
-        // Move `report_builder` from the report JSON's parse context to this one
-        let chunks_ctx = chunks::ParseCtx::new(report_builder_tx, files, sessions);
-        let mut chunks_stream =
-            chunks::ReportOutputStream::<&str, SqliteReport, SqliteReportBuilderTx> {
-                input: buf,
-                state: chunks_ctx,
-            };
-        chunks::parse_chunks_file
-            .parse_next(&mut chunks_stream)
-            .map_err(|e| e.into_inner().unwrap_or_default())
-            .map_err(CodecovError::ParserError)?;
-    }
+    chunks::parse_chunks_file(&chunks_file, report_json, report_builder_tx)?;
 
     Ok(())
 }
